[Sanity Check] No Git Repo
