+++ src/lib/utils/index.ts
@@ -1,57 +1,29 @@
-import chalk from "chalk";
-import { execSync } from "child_process";
-import fs from "fs-extra";
-import os from "os";
-import path from "path";
+import {
+  CURRENT_REPO_CONFIG_PATH,
+  makeId,
+  repoConfig,
+  trunkBranches,
+  userConfig,
+} from "./config";
+import { gpExecSync } from "./exec_sync";
+import {
+  logError,
+  logErrorAndExit,
+  logInfo,
+  logSuccess,
+  logWarn,
+} from "./splog";
 
-const CONFIG_NAME = ".graphite_repo_config";
-const USER_CONFIG_PATH = path.join(os.homedir(), CONFIG_NAME);
-type UserConfigT = {
-  branchPrefix?: string;
+export {
+  CURRENT_REPO_CONFIG_PATH,
+  makeId,
+  userConfig,
+  repoConfig,
+  trunkBranches,
+  gpExecSync,
+  logError,
+  logErrorAndExit,
+  logInfo,
+  logSuccess,
+  logWarn,
 };
-type RepoConfigT = {
-  trunkBranches?: string[];
-};
-
-export const CURRENT_REPO_CONFIG_PATH: string | undefined = (() => {
-  const repoRootPath = execSync(`git rev-parse --show-toplevel`)
-    .toString()
-    .trim();
-  if (!repoRootPath || repoRootPath.length === 0) {
-    return undefined;
-  }
-  return path.join(repoRootPath, CONFIG_NAME);
-})();
-
-export function makeId(length: number): string {
-  let result = "";
-  const characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
-  const charactersLength = characters.length;
-  for (let i = 0; i < length; i++) {
-    result += characters.charAt(Math.floor(Math.random() * charactersLength));
-  }
-  return result;
-}
-
-// TODO: validate the shape of this (possibly using retype)
-export let userConfig: UserConfigT = {};
-if (fs.existsSync(USER_CONFIG_PATH)) {
-  const userConfigRaw = fs.readFileSync(USER_CONFIG_PATH);
-  try {
-    userConfig = JSON.parse(userConfigRaw.toString().trim()) as UserConfigT;
-  } catch (e) {
-    console.log(chalk.yellow(`Warning: Malformed ${USER_CONFIG_PATH}`));
-  }
-}
-
-export let repoConfig: RepoConfigT = {};
-if (CURRENT_REPO_CONFIG_PATH && fs.existsSync(CURRENT_REPO_CONFIG_PATH)) {
-  const repoConfigRaw = fs.readFileSync(CURRENT_REPO_CONFIG_PATH);
-  try {
-    repoConfig = JSON.parse(repoConfigRaw.toString().trim()) as RepoConfigT;
-  } catch (e) {
-    console.log(chalk.yellow(`Warning: Malformed ${CURRENT_REPO_CONFIG_PATH}`));
-  }
-}
-
-export const trunkBranches: string[] | undefined = repoConfig.trunkBranches;
